Bugfix: il casting EDOOOOOOOOOOO!!!!! mapVolume in class_map and getVolume in class_octant_2D corrected

+++ include/Class_Map.hpp
@@ -122,7 +122,7 @@ public:
 	 * \param[in] volume Volume of octant from logical domain.
 	 * \return Coordinate Volume of octant in physical domain.
 	 */
-	double mapVolume(uint32_t const & volume);
+	double mapVolume(uint64_t const & volume);
 
 	/*! Transformation of coordinates of center of an octant.
 	 * \param[in] center Pointer to coordinates of center from logical domain.
